cli/tests/projects/auth-example/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
requests==2.32.3
azure-mgmt-resource==23.1.1
azure-mgmt-storage==21.2.1
azure-core==1.32.0
